Win32 warning fixes

+++ examples/platform/vsgwin32/vsgwin32.cpp
@@ -105,7 +105,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPreInst, LPSTR lpszCmdLine, i
     thread->join();
     delete thread;
 
-    return (msg.wParam);
+    return static_cast<int>(msg.wParam);
 }
 
 LRESULT CALLBACK WndProc(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam)
